new version of KDTree added, KMeans initialization has been improved; a few other useful methods and functions have been added
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@1870 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ src/cvaux/cvocttree.cpp
@@ -267,7 +267,7 @@ void OctTree::buildTree( const Vector<Point3f>& points3d, int maxLevels, int min
     root.isLeaf = true;
     root.maxLevels = maxLevels;
     root.begin = 0;
-    root.end = points.size();
+    root.end = (int)points.size();
     for( int i = 0; i < 8; i++ )
         root.children[i] = 0;
 
@@ -322,8 +322,8 @@ void  OctTree::buildNext(size_t node_ind)
 
         child.isLeaf = true;
         child.maxLevels = nodes[node_ind].maxLevels - 1;		
-        child.begin = nodes[node_ind].begin + boxBorders[i+0];
-        child.end   = nodes[node_ind].begin + boxBorders[i+1];
+        child.begin = nodes[node_ind].begin + (int)boxBorders[i+0];
+        child.end   = nodes[node_ind].begin + (int)boxBorders[i+1];
         for( int k = 0; k < 8; k++ )
             child.children[k] = 0;
